feat(site): app-router home with build-time rss + sitemap side effects
